Validate form in api action
