+++ Internal/XTI_Tool.Extensions/XTI_Tool.Extensions.csproj
@@ -10,7 +10,7 @@
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
     <PackageReference Include="MainDB.EF.SqlServer" Version="1.9.4" />
-    <PackageReference Include="XTI_Secrets.Extensions" Version="1.7.3" />
+    <PackageReference Include="XTI_Secrets.Extensions" Version="1.7.4" />
   </ItemGroup>
 
   <ItemGroup>
