+++ Tests/XTI_App.Tests/XTI_App.Tests.csproj
@@ -8,12 +8,12 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
-    <PackageReference Include="NUnit" Version="3.12.0" />
+    <PackageReference Include="NUnit" Version="3.13.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
     <PackageReference Include="MainDB.EF.InMemory" Version="1.9.4" />
     <PackageReference Include="XTI_Core.Fakes" Version="1.8.9" />
-    <PackageReference Include="XTI_Secrets.Fakes" Version="1.7.3" />
+    <PackageReference Include="XTI_Secrets.Fakes" Version="1.7.4" />
   </ItemGroup>
 
   <ItemGroup>
